requirements.txt
Flask==3.0.0
python-telegram-bot==20.0
google-api-python-client==2.91.0
google-auth==2.23.0
google-auth-oauthlib==1.2.0
requests==2.31.0
